refactor(pauli_string_collection/get_algebra_basis): delegate get_algebra_basis to Classification
