+++ src/paulie/classifier/types.py
@@ -1,0 +1,15 @@
+"""
+Common types for classifier
+"""
+
+import enum
+
+class TypeAlgebra(enum.Enum):
+    """
+    Lie algebra type.
+    """
+
+    U = 0
+    SU = 1
+    SP = 2
+    SO = 3
