+++ src/paulie/common/pauli_string_collection.py
@@ -13,7 +13,6 @@ import networkx as nx
 from paulie.common.pauli_string_bitarray import PauliString
 from paulie.common.pauli_string_linear import PauliStringLinear
 from paulie.common.get_graph import get_graph
-from paulie.common.algebra_basis import get_algebras_basis
 from paulie.classifier.classification import Classification
 from paulie.classifier.tracked_canonicalizer import TrackedCanonicalizer
 from paulie.classifier.canonicalizer import Canonicalizer
@@ -604,10 +603,9 @@ class PauliStringCollection:
         return bool(classification.is_algebra(algebra))
 
     def get_algebra_basis(self) -> np.ndarray:
-        """Constructs and returns the basis matrices for the algebras present in the classification.
-
-        Aggregates algebra properties (type, size, and multiplicity) from the
-        morphological classification and generates a combined basis representation.
+        """
+        Get the matrix basis of the dynamical Lie algebra generated by 
+        the set of Pauli strings in its defining representation.
 
         Returns:
             np.ndarray: A 3D array of shape (dim, n, n) containing the concatenated
@@ -617,28 +615,7 @@ class PauliStringCollection:
             ValueError: If the identified algebra size is invalid for basis generation.
             TypeError: If the classification properties are not properly initialized.
         """
-        classification = self.get_class()
-        morphs = classification.get_morphs()
-
-        # Key: Tuple of (TypeAlgebra, size), Value: Total multiplier/dimension count
-        algebra_map = {}
-        for morph in morphs:
-            type_alg, nc, size = morph.get_algebra_properties()
-            multiplier = nc if nc == 1 else 2 ** (nc - 1)
-            key = (type_alg, size)
-            algebra_map[key] = algebra_map.get(key, 0) + multiplier
-
-        # Prepare the input lists for basis generator
-        multipliers = []
-        groups = []
-        sizes = []
-
-        for (type_alg, size), count in algebra_map.items():
-            multipliers.append(count)
-            groups.append(type_alg)
-            sizes.append(size)
-
-        return get_algebras_basis(multipliers, groups, sizes)
+        return self.get_class().get_algebra_basis()
 
     def gen_generators(self) -> Generator[PauliStringCollection, None, None]:
         """
